backend/pyproject.toml
[project]
name = "market-scanner"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.26",
  "scikit-learn>=1.3",
  "tensorflow>=2.13",
  "yfinance>=0.2.40",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "pyyaml>=6.0"
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
